add admin check to build.ps1 script

+++ build.ps1
@@ -1,4 +1,14 @@
+function Is-Admin() {
+    $currentPrincipal = New-Object Security.Principal.WindowsPrincipal([Security.Principal.WindowsIdentity]::GetCurrent())
+    return $currentPrincipal.IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
+}
+
 function main() {
+    if (-not (Is-Admin)) {
+        Write-Host "error: administrator privileges required"
+        return 1
+    }
+
     if (Test-Path ".\build\") {
         Remove-Item -Path ".\build\" -Recurse -Force
     }
